feat: suspense 경계 세분화
