+++ apps/client/src/app/[user]/components/LeftSidebar/SidebarProvider/hook.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { SidebarContext } from ".";
+
+export const useSidebrChip = () => {
+  const ctx = useContext(SidebarContext);
+  if (!ctx) {
+    throw new Error("useSidebarChip must be used within a SidebarProvider");
+  }
+  return ctx;
+};
