+++ apps/client/src/app/[user]/components/LeftSidebar/StudyCount.tsx
@@ -1,0 +1,22 @@
+import { useMyGroupsQueryObject } from "@/app/api/users/query";
+import { prefetchQuery } from "@/shared/util/prefetch";
+import { countTextStyle, countWrapper } from "./index.css";
+
+const StudyCount = async () => {
+  const queryOption = useMyGroupsQueryObject();
+  const queryClient = await prefetchQuery(queryOption);
+  const myGroups = queryClient.getQueryData(queryOption.queryKey);
+  await new Promise((resolve) => setTimeout(resolve, 5000));
+
+  const studyCount = myGroups
+    ? Object.values(myGroups).reduce((acc, val) => acc + val.length, 0)
+    : 0;
+
+  return (
+    <div className={countWrapper}>
+      <span className={countTextStyle}>{studyCount}</span>
+    </div>
+  );
+};
+
+export default StudyCount;
